+++ views/data.ejs
@@ -12,17 +12,17 @@
         <!-- Header Section -->
         <header class="d-flex flex-md-row justify-content-between align-items-start align-items-md-center mb-4">
             <h1 class="h3 mb-3 mb-md-0">Book Data</h1>
-            <div class="d-flex flex-column flex-md-row gap-2"> <!-- Added flex-column for mobile screens -->
+            <div class="d-flex flex-column flex-md-row gap-2"> 
                 <a href="/dashboard" class="btn btn-dark mb-2 mb-md-0">Add Book</a>
                 <a href="/logout" class="btn btn-dark">Logout</a>
             </div>
         </header>
         
-        
-        <!-- Error Message -->
+    
+      
         <div class="alert alert-danger d-none" id="error-message" role="alert"></div>
 
-        <!-- Book Table (Responsive) -->
+       
         <div class="table-responsive">
             <table class="table table-striped table-hover shadow-sm" id="book-table">
                 <thead class="table-dark">
@@ -34,9 +34,7 @@
                         <th>Actions</th>
                     </tr>
                 </thead>
-                <tbody>
-                    <!-- Dynamic rows will be added here -->
-                </tbody>
+               
             </table>
         </div>
     </div>
