+++ views/forgot-password.ejs
@@ -19,7 +19,7 @@
                 <button type="submit" class="btn btn-dark w-100">Send</button>
             </form>
             <p class="text-center mt-3">
-                Remembered your password? <a href="/login">Login</a>
+                Remember your password? <a href="/login">Login</a>
             </p>
             <% if (message) { %>
             <p class="text-center text-danger"><%= message %></p>
