+++ views/profile.ejs
@@ -1,9 +1,10 @@
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Book Data</title>
+    <title>My Profile</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
@@ -12,24 +13,28 @@
             <h1 class="h3 mb-3 mb-md-0">My Profile</h1>
             <div class="d-flex flex-column flex-md-row gap-2">
                 <a href="/data" class="btn btn-dark mb-2 mb-md-0">Dashboard</a>
-                <a href="/logout" class="btn btn-dark">Logout</a>
+                <a href="/login" class="btn btn-dark">Logout</a>
             </div>
         </header>
-        <div class="alert alert-danger d-none" id="error-message" role="alert"></div>
         <div class="table-responsive">
             <table class="table table-striped table-hover shadow-sm" id="book-table">
                 <thead class="table-dark">
                     <tr>
-                        <th>Name</th>
-                        <th>Username</th>
+                        <th>Email</th>
+                        
                         <th>Actions</th>
                     </tr>
                 </thead>
-                
+                <tbody>
+                    <tr>
+                        <td><%= user.email %></td> 
+                        <td>
+                            <a href="/login" class="btn btn-primary btn-sm">Edit</a>
+                            </td>
+                    </tr>
+                </tbody>
             </table>
         </div>
     </div>
-
-    <script type="module" src="my-account.js"></script>
 </body>
 </html>
